Safari-609.1.11.1/LayoutTests/editing/selection/home-inside-noneditable-table.html

<script src="../../resources/dump-as-markup.js"></script>
<table>
  <tr>
    <td>
    <div id="test" contenteditable="true">
      <div id="l1">The caret is initially here but should move to the beginning of the line.</div>
      <div>dummy text</div>
    </div>
    </td>
  </tr>
</table>
<script>

Markup.description("Tests whether home moves the caret to the beginning of line inside a content-editable in an uneditable table.")

var s = window.getSelection();
s.setPosition(document.getElementById('l1').firstChild, 23);
Markup.dump('test');
s.modify("move", "backward", "lineboundary");
Markup.dump('test');

</script>
